# tools/ui_server/openapi.yaml
openapi: 3.0.3
info:
  title: Shankpit UI Server
  version: "0.1.0"
servers:
  - url: http://localhost:17777
paths:
...
  /api/v0/ui/intent:
    post:
      summary: Send a UI intent (HTTP fallback; WS preferred)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntentEnvelope"
      responses:
        "200":
          description: Intent accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntentResult"
...
  /api/v0/world/travel:
    post:
      summary: Request travel to a scene
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelRequest"
      responses:
        "200":
          description: travel result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntentResult"
...
components:
  schemas:
...
    IntentEnvelope:
      type: object
      required: [protocolVersion, intentId, type, ts]
      properties:
        protocolVersion: { type: string, example: "v0" }
        intentId: { type: string }
        type:
          type: string
          enum:
            - intent.activate
            - intent.open
            - intent.back
            - intent.editLabel
            - intent.select
            - intent.portalTravel
            - intent.enterVehicle
        ts: { type: integer, description: "ms since epoch or monotonic" }
        payload:
          type: object
          additionalProperties: true
    IntentResult:
      type: object
      required: [intentId, ok]
      properties:
        intentId: { type: string }
        ok: { type: boolean }
        message: { type: string }
        patch:
          type: array
          description: "Optional JSON Patch to apply to UiState"
          items: { type: object }
...
    TravelRequest:
      type: object
      required: [sceneId]
      properties:
        sceneId: { type: string }